BT_Subtree_Insert: fix interesting bug
When calling DB_Get_NextBlock() we passed int* to size_t*. This caused a
cast from int* to size_t* inside the scope of DB_Get_NextBlock(). When
memcpy ran it overflowed destination because size_t* is pointing to a
memory with sizeof(int) = 4 bytes available and was writing
sizeof(size_t) = 8 bytes overflowing the memory.

This overflow caused _probably_ an overwrite over the program's stack
bytes, changing the value of the stored BF_Block* block, making it
NULL/invalid, causing the rest of the execution to fail without an
apparent reason.

Moral of the story: don't cast pointers to different pointer type of
smaller type memory.
i.e.: Casting from int* to size_t* is dangerous because size_t is 8 bytes
and int is 4 bytes. Be careful when doing so and have full knowledge of
the related risks, like memcpy in this example.

+++ src/index/BT.c
@@ -66,7 +66,7 @@ int BT_Subtree_Insert(int file_desc_AM, int subtree_root, Record record,
 	int fieldA_length;
 	int fieldB_length;
 	size_t n_entries;  // Maximum number of entries in a data block.
-	int next_block;
+	size_t next_block;
 	/********************************************************/
 
 	int next_subtree_root;  // The root of the next block when descending.
